+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_last_successful_sync.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import companysyncstatus as shared_companysyncstatus
-from ..shared import errormessage as shared_errormessage
+from ...models.shared import companysyncstatus as shared_companysyncstatus
 from typing import Optional
 
 
@@ -22,11 +21,9 @@ class GetLastSuccessfulSyncResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     company_sync_status: Optional[shared_companysyncstatus.CompanySyncStatus] = dataclasses.field(default=None)
     r"""Success"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""