+++ sync-for-commerce/src/codatsynccommerce/models/operations/get_sync_flow_url.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import syncflowurl as shared_syncflowurl
+from ...models.shared import syncflowurl as shared_syncflowurl
 from typing import Optional
 
 
@@ -25,7 +25,7 @@ class GetSyncFlowURLResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
+    raw_response: requests_http.Response = dataclasses.field()
     r"""Raw HTTP response; suitable for custom response parsing"""
     sync_flow_url: Optional[shared_syncflowurl.SyncFlowURL] = dataclasses.field(default=None)
     r"""Success"""