+++ sync-for-commerce/src/codatsynccommerce/models/operations/set_configuration.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import configuration as shared_configuration
+from ...models.shared import configuration as shared_configuration
 from typing import Optional
 
 
@@ -21,9 +21,9 @@ class SetConfigurationResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     configuration: Optional[shared_configuration.Configuration] = dataclasses.field(default=None)
     r"""Success"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""