+++ sync-for-commerce/src/codatsynccommerce/models/shared/__init__.py
@@ -17,7 +17,8 @@ from .customer import *
 from .dataconnectionerror import *
 from .dataconnectionstatus import *
 from .datatypefeature import *
-from .errormessage import *
+from .errorvalidation import *
+from .errorvalidationitem import *
 from .featurestate import *
 from .featuretype import *
 from .fees import *
@@ -32,7 +33,9 @@ from .integrations import *
 from .invoicelevelselection import *
 from .invoicelinelevelselection import *
 from .invoicestatus import *
+from .items import *
 from .links import *
+from .locale import *
 from .localization import *
 from .newpayments import *
 from .newtaxrates import *
@@ -53,4 +56,4 @@ from .taxratemapping import *
 from .updateconnection import *
 from .visibleaccounts import *
 
-__all__ = ["AccountOption","Branding","BrandingButton","BrandingImage","BrandingLogo","Companies","Company","CompanySyncStatus","ConfigAccount","Configuration","Connection","ConnectionSourceType","Connections","CreateCompany","Customer","DataConnectionError","DataConnectionStatus","DataTypeFeature","DataTypeFeatureDataTypes","ErrorMessage","FeatureState","FeatureType","Fees","FeesSupplier","Grouping","GroupingLevels","GroupingPeriod","HalRef","ImageReference","Integration","Integrations","InvoiceLevelSelection","InvoiceLineLevelSelection","InvoiceStatus","Links","Localization","NewPayments","NewTaxRates","Option","Payments","Sales","Security","SourceType","SupportedFeature","SyncConnectionDeletedWebhook","SyncFlowURL","SyncRange","SyncRangeDateRange","SyncStatus","SyncSummary","SyncSummarySyncDateRangeUtc","SyncToLatestArgs","TaxRateAmount","TaxRateMapping","UpdateConnection","VisibleAccounts"]
+__all__ = ["AccountOption","Branding","BrandingButton","BrandingImage","BrandingLogo","Companies","Company","CompanySyncStatus","ConfigAccount","Configuration","Connection","Connections","CreateCompany","Customer","DataConnectionError","DataConnectionStatus","DataTypeFeature","DataTypes","DateRange","ErrorValidation","ErrorValidationItem","FeatureState","FeatureType","Fees","FeesSupplier","GroupReference","Grouping","GroupingLevels","GroupingPeriod","HalRef","ImageReference","Integration","Integrations","InvoiceLevelSelection","InvoiceLineLevelSelection","InvoiceStatus","Items","Links","Locale","Localization","NewPayments","NewTaxRates","Option","Payments","Sales","Security","SourceType","SupportedFeature","SyncConnectionDeletedWebhook","SyncDateRangeUtc","SyncFlowURL","SyncRange","SyncStatus","SyncSummary","SyncToLatestArgs","TaxRateAmount","TaxRateMapping","UpdateConnection","VisibleAccounts"]