+++ sync-for-commerce/src/codatsynccommerce/models/shared/configaccount.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import accountoption as shared_accountoption
+from .accountoption import AccountOption
 from codatsynccommerce import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import List, Optional
@@ -12,15 +12,16 @@ from typing import List, Optional
 @dataclasses.dataclass
 class ConfigAccount:
     r"""G/L account object for configuration."""
-    account_options: Optional[List[shared_accountoption.AccountOption]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('accountOptions') }})
+    UNSET='__SPEAKEASY_UNSET__'
+    account_options: Optional[List[AccountOption]] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('accountOptions'), 'exclude': lambda f: f is ConfigAccount.UNSET }})
     r"""Object containing account options."""
-    description_text: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('descriptionText') }})
+    description_text: Optional[str] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('descriptionText'), 'exclude': lambda f: f is ConfigAccount.UNSET }})
     r"""Descriprtive text for sales configuration section."""
-    label_text: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('labelText') }})
+    label_text: Optional[str] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('labelText'), 'exclude': lambda f: f is ConfigAccount.UNSET }})
     r"""Label text for sales configuration section."""
     required: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('required'), 'exclude': lambda f: f is None }})
     r"""Required section to be configured for sync."""
-    selected_account_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedAccountId') }})
+    selected_account_id: Optional[str] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedAccountId'), 'exclude': lambda f: f is ConfigAccount.UNSET }})
     r"""Selected account id from the list of available accounts."""