+++ sync-for-commerce/src/codatsynccommerce/models/shared/customer.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 import dataclasses
-from ..shared import option as shared_option
+from .option import Option
 from codatsynccommerce import utils
 from dataclasses_json import Undefined, dataclass_json
 from typing import List, Optional
@@ -11,9 +11,10 @@ from typing import List, Optional
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class Customer:
-    customer_options: Optional[List[shared_option.Option]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('customerOptions') }})
+    UNSET='__SPEAKEASY_UNSET__'
+    customer_options: Optional[List[Option]] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('customerOptions'), 'exclude': lambda f: f is Customer.UNSET }})
     r"""List of customer options from the list of customer records on the accounting software."""
-    selected_customer_id: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedCustomerId') }})
+    selected_customer_id: Optional[str] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedCustomerId'), 'exclude': lambda f: f is Customer.UNSET }})
     r"""Selected customer id from the list of customer records on the accounting software."""