+++ sync-for-commerce/src/codatsynccommerce/models/shared/invoicestatus.py
@@ -10,9 +10,10 @@ from typing import List, Optional
 @dataclass_json(undefined=Undefined.EXCLUDE)
 @dataclasses.dataclass
 class InvoiceStatus:
-    invoice_status_options: Optional[List[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('invoiceStatusOptions') }})
+    UNSET='__SPEAKEASY_UNSET__'
+    invoice_status_options: Optional[List[str]] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('invoiceStatusOptions'), 'exclude': lambda f: f is InvoiceStatus.UNSET }})
     r"""Options for invoice statuses."""
-    selected_invoice_status: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedInvoiceStatus') }})
+    selected_invoice_status: Optional[str] = dataclasses.field(default=UNSET, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('selectedInvoiceStatus'), 'exclude': lambda f: f is InvoiceStatus.UNSET }})
     r"""Selected option for invoice status for invoice to be synced."""